DFT/SpectralWindow UnitTest is now in a separate file.
